Rewritten task_info
